s/sound: move sound routines from sndpc
