+++ src/game/sound.c
@@ -4,7 +4,7 @@
 #include "game/game.h"
 #include "global/vars.h"
 #include "specific/init.h"
-#include "specific/sndpc.h"
+#include "specific/sound.h"
 #include "util.h"
 
 #include <stddef.h>
