+++ src/specific/sndpc.h
@@ -6,17 +6,6 @@
 
 #include "global/types.h"
 
-int32_t ConvertVolumeToDecibel(int32_t volume);
-int32_t ConvertPanToDecibel(uint16_t pan);
-
-bool SoundInit();
-void SoundLoadSamples(char **sample_pointers, int32_t num_samples);
-SAMPLE_DATA *SoundLoadSample(char *content);
-int32_t SoundMakeSample(SAMPLE_DATA *sample_data);
-void *SoundPlaySample(
-    int32_t sample_id, int32_t volume, int16_t pitch, uint16_t pan,
-    int8_t loop);
-
 int32_t MusicInit();
 int32_t MusicPlay(int16_t track_id);
 void MusicPlayLooped();
@@ -26,14 +15,5 @@ void S_MusicLoop();
 void S_MusicVolume(int16_t volume);
 void S_MusicPause();
 void S_MusicUnpause();
-
-void *S_SoundPlaySample(
-    int32_t sample_id, uint16_t volume, uint16_t pitch, int16_t pan);
-void *S_SoundPlaySampleLooped(
-    int32_t sample_id, uint16_t volume, uint16_t pitch, int16_t pan);
-int32_t S_SoundSampleIsPlaying(void *handle);
-void S_SoundStopAllSamples();
-void S_SoundStopSample(void *handle);
-void S_SoundSetPanAndVolume(void *handle, int16_t pan, int16_t volume);
 
 #endif
